+++ data/duke.txt
@@ -1,2 +1,2 @@
-T,1,read book
-D,1,something,5pm
+T,1,something else
+E,1,something else,2001-09-09 20:30
